requirements.txt
sqlite3
ultralytics
opencv-python
Pillow
arabic_reshaper
python-bidi
numpy